% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-group-limits.R
\name{add_group_limits}
\alias{add_group_limits}
\title{Add minimum and maximum values for columns by groups}
\usage{
add_group_limits(df, col1, col2, .by)
}
\arguments{
\item{df}{A data frame.}

\item{col1}{A column name for which to calculate the minimum value.}

\item{col2}{A column name for which to calculate the maximum value.}

\item{.by}{A single grouping variable, or a list of grouping variables.}
}
\value{
A data frame with two new columns: one containing the minimum value
of col1 for each group, and the other containing the maximum value of col2
for each group
}
\description{
This function adds two new columns to a data frame. The first column contains
the minimum value of a specified column for each group, and the second column
contains the maximum value of another specified column for each group.
}
