% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-loci.R
\name{build_loci_wo_pseudos}
\alias{build_loci_wo_pseudos}
\title{Build loci, removing pseudogenes in the process}
\usage{
build_loci_wo_pseudos(padlocout, index, path_db, distance)
}
\description{
Build loci, removing pseudogenes in the process
}
