% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cath.R
\name{cath_resolve_hits_process}
\alias{cath_resolve_hits_process}
\title{Process a summarised \code{cath-resolve-hits} table}
\usage{
cath_resolve_hits_process(cath_resolve_hits_summarised)
}
\arguments{
\item{cath_resolve_hits_summarised}{As read-in by \code{\link[=read_cath_resolve_hits_summarised]{read_cath_resolve_hits_summarised()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Process a summarised \code{cath-resolve-hits} table.
}
