% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{collapse_padlocout}
\alias{collapse_padlocout}
\title{Collapse a PADLOC output table to one row per system}
\usage{
collapse_padlocout(padlocout, select = NULL, drop = FALSE)
}
\arguments{
\item{padlocout}{A PADLOC output table (as read-in by \code{\link[=read_padlocout]{read_padlocout()}}).}

\item{select}{Optionally, a selection of columns to}

\item{drop}{Relevant when \code{select} is specified. If \code{drop = FALSE} (the
default), the output table retains only the columns specified. If
\code{drop = TRUE}, column selection is inversed. Selection occurs before
collapsing, so dropping uneeded columns may speed up processing.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated as the same functionality can be achieved by
using \code{\link[tidyr:nest]{tidyr::nest()}}.
Collapse a PADLOC output table to one row per system.
}
