% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gff.R
\name{combine_gff}
\alias{combine_gff}
\title{Combine a list of GFF tables}
\usage{
combine_gff(gff_list)
}
\arguments{
\item{gff_list}{A list of GFF tables (as read in by \code{\link[=multi_read_gff]{multi_read_gff()}}).}
}
\value{
A combined GFF table, introducing a new row \code{genome_accession},
which is derived from the names of the GFF tables in \code{padlocout_list}.
}
\description{
Combine a list of GFF tables
}
