% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{combine_padlocout}
\alias{combine_padlocout}
\title{Combine a list of padlocout tables}
\usage{
combine_padlocout(padlocout_list)
}
\arguments{
\item{padlocout_list}{A list of padlocout tables (as read in by
[multi_read_padlocout()]).}
}
\value{
A combined padlocout table, introducing a new row `genome_accession`,
which is derived from the names of the padlocout tables in `padlocout_list`.
}
\description{
Combine a list of padlocout tables
}
