% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{expand_padlocout}
\alias{expand_padlocout}
\title{Expand a collapsed PADLOC output table}
\usage{
expand_padlocout(collapsed_padlocout)
}
\arguments{
\item{collapsed_padlocout}{A collapsed PADLOC output table}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated as the same functionality can be achieved by
using \code{\link[tidyr:unnest]{tidyr::unnest()}}.
}
