% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-clusters.R
\name{extend_clusters}
\alias{extend_clusters}
\title{Assign second order cluster identities to any first order clusters that did
not cluster with each other.}
\usage{
extend_clusters(first_order_clusters, second_order_clusters)
}
\arguments{
\item{first_order_clusters}{A table of first order clusters, such as that
generated by MMseqs2 and read-in w/ \link{read_mmseqs_cluster_tsv}.}

\item{second_order_clusters}{A table of second order clusters, such as that
generated from an HHcat graph}
}
\description{
Both the \code{first_order_clusters} and \code{second_order_clusters} tables should
have two columns, \code{clu_member} and \code{clu_representative}, where \code{clu_member}
contains the ID's of all elements that were clustered, and where
\code{clu_representative} contains the ID's of all resulting clusters - grouping
the respective elements together. The \code{clu_member} column of the
\code{second_order_clusters} should contain elements from the \code{clu_representative}
column of the \code{first_order_clusters}.
}
