% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_extract.R
\name{future_seq_extract}
\alias{future_seq_extract}
\title{Parallel implementation of \code{\link[=seq_extract]{seq_extract()}} using \code{{furrr}}}
\usage{
future_seq_extract(path_db, seq_extract_index)
}
\arguments{
\item{path_db}{The path to use as a prefix for the \code{path} column of
\code{seq_extract_index}.}

\item{seq_extract_index}{A data frame with two columns, \code{id} and \code{path}, as
created by \code{make_seq_extract_index}. Where \code{id} contains the sequence
identifiers for extraction and \code{path} contains the paths to the respective
FASTA files (relative to \code{path_db}).}
}
\value{
A list of character vectors containing sequence information.
}
\description{
Parallel implementation of \code{\link[=seq_extract]{seq_extract()}} using \code{{furrr}}
}
