% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{gff_mutate_relative_position}
\alias{gff_mutate_relative_position}
\title{Create a column for relative position}
\usage{
gff_mutate_relative_position(gff, name = "relative.position")
}
\description{
Create a column for relative position
}
