% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-molecular-units.R
\name{group_by_distance}
\alias{group_by_distance}
\title{Assign rows to groups based on two variables describing position
(e.g. start and end).}
\usage{
group_by_distance(df, col1, col2, n, name, .by)
}
\arguments{
\item{df}{A data frame.}

\item{col1}{The column containing start position information}

\item{col2}{The column containing end position information}

\item{n}{The maximum difference between positions for grouping.}

\item{name}{The name of the new grouping column.}

\item{by}{Columns to group by.}
}
\description{
For example, grouping genetic units such as genes or systems based on the
number of nucleotides between where one unit ends and the next begins.
}
