% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-molecular-units.R
\name{group_by_position}
\alias{group_by_position}
\title{Assign rows to groups based on a single variable describing position}
\usage{
group_by_position(df, col, n, name, .by)
}
\arguments{
\item{df}{A data frame.}

\item{col}{The column containing position information.}

\item{n}{The maximum difference between positions for grouping.}

\item{name}{The name of the new grouping column.}

\item{by}{Columns to group by.}
}
\description{
For example, grouping genetic units such as genes or systems based on their
relative position in the genome.
}
