% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-hhsearch.R
\name{hhsearch_hhprob_process}
\alias{hhsearch_hhprob_process}
\title{Process a HHsearch hhprob table}
\usage{
hhsearch_hhprob_process(hhsearch_hhprob)
}
\arguments{
\item{hhsearch_hhprob}{As read-in by \code{\link[=read_hhsearch_hhprob]{read_hhsearch_hhprob()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Process a HHsearch hhprob table.
}
