% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_extract.R
\name{make_seq_extract_index}
\alias{make_seq_extract_index}
\title{Prepare an index table for parallel sequence extraction}
\usage{
make_seq_extract_index(loci_tbl, index, col_id, col_path)
}
\arguments{
\item{loci_tbl}{A loci table, as prepared with \code{\link[=build_loci]{build_loci()}} (and probably
post-processed with \code{\link[=gff_separate_attributes]{gff_separate_attributes()}}.}

\item{index}{An index table with at least a column \code{genome.accession}, with
the names of the genomes that appear in the \code{loci_tbl} anda column
with the respective paths to each genome FASTA file.}

\item{col_id}{The column in the \code{loci_tbl} which contains the sequence ID's
of interest (in most cases this is probably the column \code{Name}, which can be
pulled out of the column \code{attributes}, with \code{\link[=gff_separate_attributes]{gff_separate_attributes()}}).}

\item{col_path}{The column in the \code{index} table which contains the paths to
the FASTA files for each genome.}
}
\value{
A data frame with two columns, \code{id} and \code{path}.
}
\description{
Basic helper function to pull all of the ID's and paths from a loci table
and accompanying index table for use with \code{\link[=future_seq_extract]{future_seq_extract()}}.
}
