% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multi_read}
\alias{multi_read}
\title{Read multiple files}
\usage{
multi_read(path = ".", func, pattern = NULL, namefix = NULL, barname)
}
\arguments{
\item{path}{A character vector of full path names; the default corresponds
to the working directory, \code{\link[=getwd]{getwd()}}. Tilde expansion (see \link{path.expand}) is
performed. Missing values will be ignored. Elements with a marked encoding
will be converted to the native encoding (and if that fails, considered
non-existent).}

\item{func}{A function to use for reading.}

\item{pattern}{An optional regular expression. Only file names which match
the regular expression will be returned.}

\item{namefix}{An optional regular expression. The name of each element
in the resulting list is derived from the file names. This regex is used to
remove a string from each file name before naming the elements.}

\item{barname}{The name of the progress bar to display.}
}
\value{
A \code{\link[=list]{list()}}, where each element is a \code{\link[tibble:tibble]{tibble::tibble()}} holding
the contents of the file.
}
\description{
Read multiple files
}
