% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{mutate_namecol}
\alias{mutate_namecol}
\title{Create a new column based on \code{\link[=names]{names()}}}
\usage{
mutate_namecol(df_list, colname, ...)
}
\arguments{
\item{df_list}{A \code{\link[=list]{list()}} of \code{\link[=data.frame]{data.frame()}}s.}

\item{colname}{The name of the new column.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Pass additional arguments to
\code{\link[dplyr:mutate]{dplyr::mutate()}} to create, modify, and delete other columns and control
where the columns should appear (i.e. \code{.before} and \code{.after}).}
}
\value{
A \code{\link[=list]{list()}} of \code{\link[=data.frame]{data.frame()}}s.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated as the same functionality can be achieved by
using the \code{.id} argument of \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}.
\code{mutate_namecol()} creates a new column for each data frame in a list,
derived from the name of the data frame.
}
\keyword{internal}
