% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padlocout-filter-embedded.R
\name{padlocout_filter_embedded}
\alias{padlocout_filter_embedded}
\title{Filter padlocout for systems with embedded genes}
\usage{
padlocout_filter_embedded(padlocout)
}
\arguments{
\item{padlocout}{A padlocout output table}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Filter padlocout for systems with embedded genes
}
