% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{padlocout_mutate_locus_limits}
\alias{padlocout_mutate_locus_limits}
\title{Create columns for minimum/maximum locus limits}
\usage{
padlocout_mutate_locus_limits(padlocout, distance)
}
\description{
Create columns for minimum/maximum locus limits
}
