% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{padlocout_mutate_separation}
\alias{padlocout_mutate_separation}
\title{Create a column for total separation between the genes of a system
NA.}
\usage{
padlocout_mutate_separation(padlocout)
}
\description{
Create a column for total separation between the genes of a system
NA.
}
