% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpad-package.R
\docType{package}
\name{postpad-package}
\alias{postpad}
\alias{postpad-package}
\title{postpad: Post-processing tools for exploring PADLOC results}
\description{
Post-processing tools for exploring PADLOC results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leightonpayne/postpad}
  \item Report bugs at \url{https://github.com/leightonpayne/postpad/issues}
}

}
\author{
\strong{Maintainer}: Leighton Payne \email{leighton.payne@postgrad.otago.ac.nz} (\href{https://orcid.org/0000-0003-2305-6827}{ORCID})

}
\keyword{internal}
