% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{postpad_script}
\alias{postpad_script}
\title{Get path to padlocdev example}
\usage{
postpad_script(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\value{
A \code{\link[base:character]{base::character()}}.
}
\description{
padlocdev comes bundled with a number of sample files in its
\code{inst/extdata} directory. This function make them easy to access.
}
