% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cath.R
\name{read_cath_resolve_hits_summarised}
\alias{read_cath_resolve_hits_summarised}
\title{Read a summarised table from \code{cath-resolve-hits}}
\usage{
read_cath_resolve_hits_summarised(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read a summarised table from \code{cath-resolve-hits}.
}
