% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-db-index.R
\name{read_db_index}
\alias{read_db_index}
\title{Read a RefSeq database index file}
\usage{
read_db_index(path)
}
\arguments{
\item{path}{The path to the RefSeq database index.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
A RefSeq database has RefSeq genome FASTA and GFF files divided into
directory 'chunks'. The index file for the database should be a TSV file with
a header and one row per genome, with columns listing the directory/chunk
where the respective genome files are located i.e. \code{chunk}, the accession
number of the genome i.e. \code{genome.accession}, i.e. how the genome is
identified in a \code{padlocout_master} table, the path to the respective genome
FASTA file i.e. \code{path_faa}, relative to the base directory of the RefSeq
database, and the path to the respective genome GFF file i.e. \code{path_gff},
relative to the base directory of the RefSeq database.
}
