% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-hhsearch.R
\name{read_hhsearch_hhprob}
\alias{read_hhsearch_hhprob}
\title{Read a HHsearch hhprob table}
\usage{
read_hhsearch_hhprob(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read a HHsearch hhprob table.
}
