% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-hhsearch.R
\name{read_hhsearch_hhtbl}
\alias{read_hhsearch_hhtbl}
\title{Read a HHsearch BLAST-formatted output table}
\usage{
read_hhsearch_hhtbl(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read a HHsearch BLAST-formatted output table.
}
