% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-mmseqs.R
\name{read_mmseqs_cluster_tsv}
\alias{read_mmseqs_cluster_tsv}
\title{Read a mmseqs cluster TSV}
\usage{
read_mmseqs_cluster_tsv(path)
}
\arguments{
\item{path}{A character vector of full path names; the default corresponds
to the working directory, \code{\link[=getwd]{getwd()}}. Tilde expansion (see \link{path.expand}) is
performed. Missing values will be ignored. Elements with a marked encoding
will be converted to the native encoding (and if that fails, considered
non-existent).}
}
\description{
Read a mmseqs cluster TSV
}
