% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-padlocout.R
\name{read_padlocout}
\alias{read_padlocout}
\title{Read PADLOC output files}
\usage{
read_padlocout(path)
}
\arguments{
\item{path}{The path to a PADLOC output file (i.e. the files generated by
PADLOC ending in '*_padloc.csv').}
}
\value{
A [tibble::tibble()].
}
\description{
Read PADLOC output files
}
