% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-padlocout.R
\name{read_padlocout_master}
\alias{read_padlocout_master}
\title{Read PADLOC master output files}
\usage{
read_padlocout_master(path)
}
\arguments{
\item{path}{The path to a PADLOC master output file (as generated by
[combine_padlocout()]).}
}
\value{
A [tibble::tibble()].
}
\description{
Read PADLOC master output files
}
