% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pfam.R
\name{read_pfam_clans}
\alias{read_pfam_clans}
\title{Read a Pfam clans table}
\usage{
read_pfam_clans(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read a Pfam clans table.
}
