% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-seqkit.R
\name{read_seqkit_duplicates}
\alias{read_seqkit_duplicates}
\title{Read a seqkit duplicates file}
\usage{
read_seqkit_duplicates(path)
}
\arguments{
\item{path}{A character vector of full path names; the default corresponds
to the working directory, \code{\link[=getwd]{getwd()}}. Tilde expansion (see \link{path.expand}) is
performed. Missing values will be ignored. Elements with a marked encoding
will be converted to the native encoding (and if that fails, considered
non-existent).}
}
\description{
Read a seqkit duplicates file
}
