% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_extract.R
\name{seq_extract}
\alias{seq_extract}
\title{Extract a DNA/protein sequence from a FASTA file}
\usage{
seq_extract(id, file)
}
\arguments{
\item{id}{The identifier of the sequence for extraction.}

\item{file}{The path to the FASTA file.}
}
\value{
A character vector with one string per line of the extracted
sequence.
}
\description{
Use the system's \code{awk} implementation to extract a sequence from a
multi-FASTA file based on the identifier in the sequence headers
e.g. the identifier for:
}
\details{
">WP_002564308.1 acyl carrier protein"

is:

"WP_002564308.1"
}
