% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-padlocout.R
\name{summarise_protein_overlap}
\alias{summarise_protein_overlap}
\title{Summarise how many times specific proteins overlap}
\usage{
summarise_protein_overlap(padlocout)
}
\arguments{
\item{padlocout}{A padlocout table (as read in by \code{\link[=read_padlocout]{read_padlocout()}}) or a
combined padlocout table (as generated by \code{\link[=padlocout_combine]{padlocout_combine()}})}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Summarise how many times specific proteins overlap
}
