% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-padlocout.R
\name{summarise_system_count}
\alias{summarise_system_count}
\title{Summarise how many of each system were identified}
\usage{
summarise_system_count(padlocout)
}
\arguments{
\item{padlocout}{A padlocout table (as read in by \code{\link[=read_padlocout]{read_padlocout()}}) or a
combined padlocout table (as generated by \code{\link[=padlocout_combine]{padlocout_combine()}})}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{system} and \code{n}, where there is a
single row for each system summarising the number of times a system was
identified in the padlocout table.
}
\description{
Summarise how many of each system were identified
}
