% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gff.R
\name{write_gff}
\alias{write_gff}
\title{Write a GFF file}
\usage{
write_gff(gff)
}
\arguments{
\item{gff}{A GFF table as read in by \code{\link[=read_gff]{read_gff()}}, or a combined GFF table,
as generated by \code{\link[=combine_gff]{combine_gff()}}).}
}
\description{
Write a GFF file
}
