% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cols_2c}
\alias{cols_2c}
\title{Combine two \code{\link[readr:cols]{readr::cols()}} specifications}
\usage{
cols_2c(cols1, cols2)
}
\arguments{
\item{cols1}{A \code{\link[readr:cols]{readr::cols()}} object.}

\item{cols2}{A \code{\link[readr:cols]{readr::cols()}} object.}
}
\value{
A \code{\link[readr:cols]{readr::cols()}} object of cols1 and cols2 combined.
}
\description{
Combine two \code{\link[readr:cols]{readr::cols()}} specifications
}
