% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cols_c}
\alias{cols_c}
\title{Combine two or more \code{\link[readr:cols]{readr::cols()}} specifications}
\usage{
cols_c(cols)
}
\arguments{
\item{cols}{A \code{\link[=list]{list()}} of \code{\link[readr:cols]{readr::cols()}} objects.}
}
\value{
A \code{\link[readr:cols]{readr::cols()}} object of all cols combined.
}
\description{
Combine two or more \code{\link[readr:cols]{readr::cols()}} specifications
}
