% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-padlocout.R
\name{multi_read}
\alias{multi_read}
\title{Read multiple files}
\usage{
multi_read(path = ".", func, pattern = NULL, name)
}
\arguments{
\item{path}{A character vector of full path names; the default corresponds
to the working directory, [getwd()]. Tilde expansion (see [path.expand]) is
performed. Missing values will be ignored. Elements with a marked encoding
will be converted to the native encoding (and if that fails, considered
non-existent).}

\item{func}{A function to use for reading.}

\item{pattern}{An optional regular expression. Only file names which match
the regular expression will be returned.}

\item{name}{The name of the progress bar to display.}
}
\value{
A [list()], where each element is a [tibble::tibble()] holding
the contents of the file.
}
\description{
Read multiple files
}
