% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-padlocout.R
\name{multi_read_padlocout}
\alias{multi_read_padlocout}
\title{Read multiple PADLOC output files}
\usage{
multi_read_padlocout(path, namefix = NULL)
}
\arguments{
\item{path}{A character vector of full path names; the default corresponds
to the working directory, \code{\link[=getwd]{getwd()}}. Tilde expansion (see \link{path.expand}) is
performed. Missing values will be ignored. Elements with a marked encoding
will be converted to the native encoding (and if that fails, considered
non-existent).}

\item{namefix}{An optional regular expression. The name of each element
in the resulting list is derived from the file names. This regex is used to
remove a string from each file name before naming the elements.}
}
\value{
A \code{\link[=list]{list()}}, where each element is a \code{\link[tibble:tibble]{tibble::tibble()}} holding
the contents of the file.
}
\description{
Read multiple PADLOC output files
}
