% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-padlocout.R
\name{multi_read_padlocout}
\alias{multi_read_padlocout}
\title{Read multiple PADLOC output files}
\usage{
multi_read_padlocout(path)
}
\arguments{
\item{path}{The path to a directory containing PADLOC output files (i.e.
the files generated by PADLOC ending in '*_padloc.csv').}
}
\value{
A [list()], where each element is a [tibble::tibble()] holding
the contents of a PADLOC output file, as read by [read_padlocout()].
}
\description{
Read multiple PADLOC output files
}
