% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{summarise_system_count}
\alias{summarise_system_count}
\title{Summarise how many of each system were identified}
\usage{
summarise_system_count(padlocout)
}
\arguments{
\item{padlocout}{A padlocout table (as read in by [read_padlocout()]) or a
combined padlocout table (as generated by [combine_padlocout()])}
}
\value{
A [tibble::tibble()] with columns `system` and `n`, where there is a
single row for each system summarising the number of times a system was
identified in the padlocout table.
}
\description{
Summarise how many of each system were identified
}
